Davis_Putnam finished, added input2 for testing
It took quite a lot of debugging, but the code for davis-putnam works
on two test inputs now given by the professor.  Fixed issues with
list removals (causes list shifts and skips during iteration),
clause and atom removal, and propogation.  Also finished up the
main dp1 algorithm with the easy and hard cases.